requirements.txt
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
scipy==1.9.1
